+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -14049,15 +14049,45 @@ func (g *GreenplumLinkedService) GetLinkedService() *LinkedService {
 
 // GreenplumLinkedServiceTypeProperties - Greenplum Database linked service properties.
 type GreenplumLinkedServiceTypeProperties struct {
+	// The authentication type to use. Type: string. Only used for V2.
+	AuthenticationType *GreenplumAuthenticationType
+
+	// The time to wait (in seconds) while trying to execute a command before terminating the attempt and generating an error.
+	// Set to zero for infinity. Type: integer. Only used for V2.
+	CommandTimeout any
+
 	// An ODBC connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
 	ConnectionString any
+
+	// The time to wait (in seconds) while trying to establish a connection before terminating the attempt and generating an error.
+	// Type: integer. Only used for V2.
+	ConnectionTimeout any
+
+	// Database name for connection. Type: string. Only used for V2.
+	Database any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// Host name for connection. Type: string. Only used for V2.
+	Host any
+
+	// The Azure key vault secret reference of password in connection string. Type: string. Only used for V2.
+	Password SecretBaseClassification
+
+	// The port for the connection. Type: integer. Only used for V2.
+	Port any
+
 	// The Azure key vault secret reference of password in connection string.
 	Pwd *AzureKeyVaultSecretReference
+
+	// SSL mode for connection. Type: integer. 0: disable, 1:allow, 2: prefer, 3: require, 4: verify-ca, 5: verify-full. Type:
+	// integer. Only used for V2.
+	SSLMode any
+
+	// Username for authentication. Type: string. Only used for V2.
+	Username any
 }
 
 // GreenplumSource - A copy activity Greenplum Database source.
@@ -27191,8 +27221,8 @@ type SapOdpLinkedServiceTypeProperties struct {
 	// resultType string).
 	SncLibraryPath any
 
-	// SNC activation indicator to access the SAP server where the table is located. Must be either 0 (off) or 1 (on). Type: string
-	// (or Expression with resultType string).
+	// SNC activation flag (Boolean) to access the SAP server where the table is located. Type: boolean (or Expression with resultType
+	// boolean).
 	SncMode any
 
 	// Initiator's SNC name to access the SAP server where the table is located. Type: string (or Expression with resultType string).
@@ -27641,8 +27671,8 @@ type SapTableLinkedServiceTypeProperties struct {
 	// resultType string).
 	SncLibraryPath any
 
-	// SNC activation indicator to access the SAP server where the table is located. Must be either 0 (off) or 1 (on). Type: string
-	// (or Expression with resultType string).
+	// SNC activation flag (Boolean) to access the SAP server where the table is located. Type: boolean (or Expression with resultType
+	// boolean).
 	SncMode any
 
 	// Initiator's SNC name to access the SAP server where the table is located. Type: string (or Expression with resultType string).
@@ -28021,6 +28051,11 @@ type ScriptActivityScriptBlock struct {
 type ScriptActivityTypeProperties struct {
 	// Log settings of script activity.
 	LogSettings *ScriptActivityTypePropertiesLogSettings
+
+	// Enable to retrieve result sets from multiple SQL statements and the number of rows affected by the DML statement. Supported
+	// connector: SnowflakeV2. Type: boolean (or Expression with resultType
+	// boolean).
+	ReturnMultistatementResult any
 
 	// ScriptBlock execution timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	ScriptBlockExecutionTimeout any
@@ -31197,6 +31232,28 @@ type TeamDeskLinkedServiceTypeProperties struct {
 	UserName any
 }
 
+// TeradataImportCommand - Teradata import command settings.
+type TeradataImportCommand struct {
+	// REQUIRED; The import setting type.
+	Type *string
+
+	// Additional format options for Teradata Copy Command. The format options only applies to direct copy from CSV source. Type:
+	// key value pairs (value should be string type) (or Expression with resultType
+	// object). Example: "additionalFormatOptions": { "timeFormat": "HHhMImSSs" }
+	AdditionalFormatOptions any
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]any
+}
+
+// GetImportSettings implements the ImportSettingsClassification interface for type TeradataImportCommand.
+func (t *TeradataImportCommand) GetImportSettings() *ImportSettings {
+	return &ImportSettings{
+		AdditionalProperties: t.AdditionalProperties,
+		Type:                 t.Type,
+	}
+}
+
 // TeradataLinkedService - Linked service for Teradata data source.
 type TeradataLinkedService struct {
 	// REQUIRED; Type of linked service.
@@ -31242,18 +31299,43 @@ type TeradataLinkedServiceTypeProperties struct {
 	// AuthenticationType to be used for connection.
 	AuthenticationType *TeradataAuthenticationType
 
-	// Teradata ODBC connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
+	// The character set to use for the connection. Type: string (or Expression with resultType string). Only applied for version
+	// 2.0.
+	CharacterSet any
+
+	// Teradata ODBC connection string. Type: string, SecureString or AzureKeyVaultSecretReference. Only applied for version 1.0.
 	ConnectionString any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// The port numbers when connecting to server through HTTPS/TLS connections. Type: integer (or Expression with resultType
+	// integer). Only applied for version 2.0.
+	HTTPSPortNumber any
+
+	// The maximum size of the response buffer for SQL requests, in bytes. Type: integer. Only applied for version 2.0.
+	MaxRespSize any
+
 	// Password for authentication.
 	Password SecretBaseClassification
 
+	// The port numbers when connecting to server through non HTTPS/TLS connections. Type: integer (or Expression with resultType
+	// integer). Only used for V2. Only applied for version 2.0.
+	PortNumber any
+
+	// SSL mode for connection. Valid values including: “Disable”, “Allow”, “Prefer”, “Require”, “Verify-CA”, “Verify-Full”. Default
+	// value is “Verify-Full”. Type: string (or Expression with resultType
+	// string). Only applied for version 2.0.
+	SSLMode any
+
 	// Server name for connection. Type: string (or Expression with resultType string).
 	Server any
+
+	// Specifies whether to encrypt all communication with the Teradata database. Allowed values are 0 or 1. This setting will
+	// be ignored for HTTPS/TLS connections. Type: integer (or Expression with
+	// resultType integer). Only applied for version 2.0.
+	UseDataEncryption any
 
 	// Username for authentication. Type: string (or Expression with resultType string).
 	Username any
@@ -31272,6 +31354,50 @@ type TeradataPartitionSettings struct {
 	// The maximum value of column specified in partitionColumnName that will be used for proceeding range partitioning. Type:
 	// string (or Expression with resultType string).
 	PartitionUpperBound any
+}
+
+// TeradataSink - A copy activity Teradata sink.
+type TeradataSink struct {
+	// REQUIRED; Copy sink type.
+	Type *string
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]any
+
+	// If true, disable data store metrics collection. Default is false. Type: boolean (or Expression with resultType boolean).
+	DisableMetricsCollection any
+
+	// Teradata import settings.
+	ImportSettings *TeradataImportCommand
+
+	// The maximum concurrent connection count for the sink data store. Type: integer (or Expression with resultType integer).
+	MaxConcurrentConnections any
+
+	// Sink retry count. Type: integer (or Expression with resultType integer).
+	SinkRetryCount any
+
+	// Sink retry wait. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
+	SinkRetryWait any
+
+	// Write batch size. Type: integer (or Expression with resultType integer), minimum: 0.
+	WriteBatchSize any
+
+	// Write batch timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
+	WriteBatchTimeout any
+}
+
+// GetCopySink implements the CopySinkClassification interface for type TeradataSink.
+func (t *TeradataSink) GetCopySink() *CopySink {
+	return &CopySink{
+		AdditionalProperties:     t.AdditionalProperties,
+		DisableMetricsCollection: t.DisableMetricsCollection,
+		MaxConcurrentConnections: t.MaxConcurrentConnections,
+		SinkRetryCount:           t.SinkRetryCount,
+		SinkRetryWait:            t.SinkRetryWait,
+		Type:                     t.Type,
+		WriteBatchSize:           t.WriteBatchSize,
+		WriteBatchTimeout:        t.WriteBatchTimeout,
+	}
 }
 
 // TeradataSource - A copy activity Teradata source.